# products_openapi.json
{"openapi":"3.1.0","info":{"title":"Bio Vegetable Distributor API","description":"Mock API for managing products of a bio vegetable distributor. Includes full CRUD functionality and automatic OpenAPI documentation.","version":"1.0.0"},"paths":{"/products":{"get":{"tags":["Products"],"summary":"Get Products","description":"Retrieve all products from the distributor.","operationId":"get_products_products_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Product"},"type":"array","title":"Response Get Products Products Get"}}}}}},"post":{"tags":["Products"],"summary":"Create Product","description":"Add a new product to the catalog.","operationId":"create_product_products_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/products/{product_id}":{"get":{"tags":["Products"],"summary":"Get Product","description":"Retrieve details of a specific product by its ID.","operationId":"get_product_products__product_id__get","parameters":[{"name":"product_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Product Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Products"],"summary":"Update Product","description":"Update an existing product's information.","operationId":"update_product_products__product_id__put","parameters":[{"name":"product_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Product Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Products"],"summary":"Delete Product","description":"Remove a product from the catalog.","operationId":"delete_product_products__product_id__delete","parameters":[{"name":"product_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Product Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Category":{"type":"string","enum":["Root","Leafy","Fruit","Brassica","Tuber","Allium","Legume"],"title":"Category"},"Certification":{"type":"string","enum":["EU-Bio","Demeter","Bioland","Naturland"],"title":"Certification"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Product":{"properties":{"name":{"type":"string","title":"Name","example":"Carrot"},"category":{"$ref":"#/components/schemas/Category","example":"Root"},"price_per_unit":{"type":"number","exclusiveMinimum":0.0,"title":"Price Per Unit","example":2.5},"unit":{"$ref":"#/components/schemas/Unit","example":"kg"},"origin":{"type":"string","title":"Origin","example":"Germany, Bavaria"},"stock_quantity":{"type":"number","minimum":0.0,"title":"Stock Quantity","example":150.0},"certification":{"$ref":"#/components/schemas/Certification","example":"Demeter"},"is_seasonal":{"type":"boolean","title":"Is Seasonal","default":true},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description","example":"Fresh, crunchy organic carrots from local farmers."},"id":{"type":"string","format":"uuid","title":"Id"}},"type":"object","required":["name","category","price_per_unit","unit","origin","stock_quantity","certification"],"title":"Product"},"ProductCreate":{"properties":{"name":{"type":"string","title":"Name","example":"Carrot"},"category":{"$ref":"#/components/schemas/Category","example":"Root"},"price_per_unit":{"type":"number","exclusiveMinimum":0.0,"title":"Price Per Unit","example":2.5},"unit":{"$ref":"#/components/schemas/Unit","example":"kg"},"origin":{"type":"string","title":"Origin","example":"Germany, Bavaria"},"stock_quantity":{"type":"number","minimum":0.0,"title":"Stock Quantity","example":150.0},"certification":{"$ref":"#/components/schemas/Certification","example":"Demeter"},"is_seasonal":{"type":"boolean","title":"Is Seasonal","default":true},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description","example":"Fresh, crunchy organic carrots from local farmers."}},"type":"object","required":["name","category","price_per_unit","unit","origin","stock_quantity","certification"],"title":"ProductCreate"},"Unit":{"type":"string","enum":["kg","piece","bunch","box"],"title":"Unit"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}